Fixed compilation warnings, errors and minor details

+++ types.h
@@ -16,7 +16,7 @@
   typedef signed   char      s8;
   typedef signed   short     s16;
   typedef signed   int       s32;
-  typedef signed   long      s64;
+  typedef signed   long long s64;
   typedef unsigned char      u8;
   typedef unsigned short     u16;
   typedef unsigned int       u32;
